+++ eminus/__init__.py
@@ -14,11 +14,13 @@ from .io import (
     read,
     read_cube,
     read_json,
+    read_traj,
     read_xyz,
     write,
     write_cube,
     write_json,
     write_pdb,
+    write_traj,
     write_xyz,
 )
 from .logger import log
@@ -26,8 +28,8 @@ from .scf import RSCF, SCF, USCF
 from .version import __version__, info
 
 __all__ = ['config', 'Atoms', 'get_epsilon', 'get_psi', 'info', 'log', 'read', 'read_cube',
-           'read_json', 'read_xyz', 'RSCF', 'SCF', 'USCF', 'write', 'write_cube', 'write_json',
-           'write_pdb', 'write_xyz', '__version__']
+           'read_json', 'read_traj', 'read_xyz', 'RSCF', 'SCF', 'USCF', 'write', 'write_cube',
+           'write_json', 'write_pdb', 'write_traj', 'write_xyz', '__version__']
 
 
 def demo():