+++ eminus/io/__init__.py
@@ -4,10 +4,11 @@ from .cube import read_cube, write_cube
 from .gth import read_gth
 from .json import read_json, write_json
 from .pdb import create_pdb_str, write_pdb
+from .traj import read_traj, write_traj
 from .xyz import read_xyz, write_xyz
 
-__all__ = ['create_pdb_str', 'read', 'read_cube', 'read_gth', 'read_json', 'read_xyz', 'write',
-           'write_cube', 'write_json', 'write_pdb', 'write_xyz']
+__all__ = ['create_pdb_str', 'read', 'read_cube', 'read_gth', 'read_json', 'read_traj', 'read_xyz',
+           'write', 'write_cube', 'write_json', 'write_pdb', 'write_traj', 'write_xyz']
 
 
 def read(*args, **kwargs):
@@ -16,6 +17,8 @@ def read(*args, **kwargs):
         return read_json(*args, **kwargs)
     if args[0].endswith('.xyz'):
         return read_xyz(*args, **kwargs)
+    if args[0].endswith(('.trj', '.traj')):
+        return read_traj(*args, **kwargs)
     if args[0].endswith(('.cub', '.cube')):
         return read_cube(*args, **kwargs)
     NotImplementedError('File ending not recognized.')
@@ -28,6 +31,8 @@ def write(*args, **kwargs):
         return write_json(*args, **kwargs)
     if args[1].endswith('.xyz'):
         return write_xyz(*args, **kwargs)
+    if args[1].endswith(('.trj', '.traj')):
+        return write_traj(*args, **kwargs)
     if args[1].endswith(('.cub', '.cube')):
         return write_cube(*args, **kwargs)
     if args[1].endswith('.pdb'):